+++ plugins/dataspex/src/scittle/dataspex.cljs
@@ -1,0 +1,9 @@
+(ns scittle.dataspex
+  {:no-doc true}
+  (:require [sci.configs.cjohansen.dataspex :refer [config]]
+            [scittle.core :as scittle]))
+
+(defn init []
+  (scittle/register-plugin!
+   ::dataspex
+   config))